Backend/requirements.txt
django>=5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.4
django-filter>=25.1
mysqlclient>=2.2
Pillow>=10.0
python-decouple>=3.8
